Add make target to build and test go binaries
This commit adds a few new make targets that build and test the
codebase.

* make components
A find command is used to automatically discover all the `main.go`
files in the components directory. The list of files is then used to
generate dynamically the targets that will build the binaries using
Docker containers. A shell script generates on the fly a Dockerfile
whose first stage is meant to package the current repo in a Dockerfile,
build the Go binary and then use a second stage to package the
resulting Go binary in a distroless container. All the binaries are
named after the name of the folder they reside in. The binaries in the
components/producers folders have the `-parser` suffix added to them.

In some cases we need custom logic to build a component or extra
tooling along with the Go binary. In order to enable such cases to
easily define their own custom logic, the shell script is intelligent
enough to be able to discover Makefiles in the directories containing
the `build` or `build-extras` target. The `build` target will be
executed instead of the default one, if one is defined. If the
`build-extras` target is defined, it will be executed once the build
step is finished. Any component can define whichever of the two targets
it needs.

In all cases the Docker repository that will be used for the tag of the
generated containers, can be modified by setting the value of the
DOCKER_REPO environment variable.

In order to build the pkg/enrichment library, the SQL migration script
path was fixed so that it can be embedded.

* make test
The make test target has been expanded to trigger Go tests for all the
packages. The user can set the value of the `GO_PACKAGES` environment
variable in order to run tests for only a subset of the packages.

* make components/base/openapi_schema.json
The `components/base/openapi_schema.json` make target has been added in
order to keep file synced with `api/openapi/tekton/openapi_schema.json`

Please BUILD files have been cleaned up to remove any targets that are
now handled by the new make targets.

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>