Cleanup kustomize-component-generator BUILD files
Removes please BUILD files from the kustomize-component-generator
directory.

Adds a Makefile target that builds the kustomize-component-generator and
outputs the binary in the `bin` directory.

The kustomize-component-generator itself has been modified to expect
command line parameters for the source task YAML and the path to the
output file. If the output file is not provided, it is set to be the
`kustomization.yaml` in the same directory as the task YAML provided as
a parameter.

Finally, adds `bin` directory to gitignore.