Add make targets for building component kustomizations (Part of #62)
This commit removes all BUILD files from the components and replaces
them with two scripts in the scripts folder and make targets that
automate building all component containers. The Go binaries are first
build using a make target with the `/bin` suffix and the path of the
directory of the component. The binaries are then packaged into
containers using `docker`. Each container is tagged with a tag that is
produced based on the git tag of the commit. If the current commit is
not tagged, the container tag will be the commit SHA + how many commits
ahead of the latest tagged commit + the latest tag in the branch's
history.
For some components, extra functionality is needed because we use a
different name for the component or some extra functionality. For these
cases, we allow a special Makefile to be placed in the directory and we
expect targets with specific names to be present. If they are, the top
level Makefile will invoke them and allow them to build the component in
whatever way they want.

Changes

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>