Improve code quality for pkg/... (Part of #62)
A lot of tests in the pkg where not checking properly for errors
returned by various function calls. We also used the `assert.Nil`
function for checking for the existence of errors, instead of the
`require.NoError` which has clearer semantics and will force the tests
to stop immediately if an error has been returned.
Finally, in many tests we included the testdata in the GO code which
means that changing the data would make it look like we changed the test
code itself. The test data where moved into separate folders to make the
code clearer to read.

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>
Signed-off-by: Pavlos Tzianos <pavlos@ocurity.com>