Improve code quality for components/...
A lot of tests for our components where not checking properlu for the
existence of errors, an issue that was identified by the linters and is
now fixed. We also switched to using `require.NoError` to make sure that
tests will stop if an error is returned anywhere. Finally, in some
tests, testdata where included as variables in the code, which have now
been split into their own files for more clarity.

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>