Cleanup third_party directory shell scripts and Tekton manifests (Part of #62)
This commit is the last commit cleaning up the third_party directory.
This directory is supposed to contain any resources that are imported
from third parties as the name suggests. The only resources that we are
importing from third parties at the moment are two shell scripts with
helpers for beautifying the output of our shell scripts and the TektonCD
OpenAPI spec which is used by us to produce OpenAPI schemas for the
CRDs.

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>