Refactor lint Github action
Remove all please related lint commands and refactor Github action to
use the newly introduced lint make target.
The lint make target invokes in turn reviewdog with the proper
configuration that allows multiple tools to be executed in parallel and
output configuration. For the Github action we make sure to invoke
reviewdog only on the changes introduced in the PR to avoid having to
correct existing issues that already existed as a prerequisite of
landing a PR.

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>