Refactor format Github action (Fixes #65)
The Github action checking the format of the code has been refactored to
use the newly introduced make target which formats the code. If the
Github action discovers that the formatting tools have introduced
changes to the codebase it will fail. That way the contributor can be
notified that the code is not formatted according to the standards.

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>
Signed-off-by: Pavlos Tzianos <pavlos@ocurity.com>