Replace kustomize component generator with draconctl (Part of #91)
The kustomize component generator refactoring was unsuccessful. A lot
of time was spent reworking something that ultimately will never work
well, due to the fact that there is some flag that we are missing when
the kustomize build command that prevents the output from being the
expected one, and the JSON 6902 patches are missing some functionality
that would allow us to make them work for us, such as appending to a
list that is a field of the last element of an existing list.
This lead us to rethink the way we do thinks and start transitioning to
a more abstract way of producing our pipelines that doesn't involve
kustomize at all. That is the beginning of the draconctl tool. That tool
will eventually include other utilities that will allow users to perform
a lot of actions related to the project easily. For the time being the
first command include is the `pipeline build` command that produces a
pipeline for Tekton. In the future we could add other backends such as
simple bash scripts that run the pipelines in demo environments where
there is no Docker or K8s available, different CI/CD platforms like Argo
CI/CD, etc or Github actions.
The new command accepts a path to a directory where a kustomization file
is located listing all the bases and the components that will be added
to the pipeline. It is able to parse a Kustomization, so it's already
compatible with all our pipelines. But it allows us to implement
arbitrary logic to the construction of our pipelines, since we now have
the full context of all the tasks involved in generating a pipeline.
This commit also adds the ability to add arbitrary tags to the scan
pipeline that are passed throughout and can help a user add more
metadata to the results of the scans.

Signed-off-by: Pavlos Tzianos <pavlos.tzianos@gmail.com>