Change workspace from source-code-ws to output (Fixes #109)
For some reason the main workspace used by all our components is called
source-code-ws. However, the default workspace used by the `git-clone`
task is called `output`. This is a more appropriate name because what we
are actually doing is get the `git-clone` task to clone the source code
into the `source-code` subdirectory of the volume and then all the rest
of the components are using that path and are creating their own
subdirectories to store their results. Using a different workspace than
the one used by default from the git-clone task is not a huge problem
but it requires us to either vendorize the taks and modify it, or add
extra configuration to an already very verbose pipeline to reference
everywhere the `output` workspace as `source-code-ws`. Clearly the
optimal solution is to reduce the hacks, hence this commit that
homogenises the whole repo to just use `output` as the workspace where
everything will be store.

Signed-off-by: Pavlos Tzianos <pavlos@ocurity.com>