producer/nancy: Mark source code repo as safe (Fixes #108)
Lately, a security feature has been added in Git where if the user
running the Git process is different than the user owning the repo, Git
will throw an error. Unless the repo has been added explicitly to a list
of safe repos.
This is an issue that affects the first step of the Nancy task, because
it uses Go tooling to list the dependencies of the repo. Go itself
figures out that it's a git repo that is being processed and invokes git
which in turn fails, because the process is running by the root user by
the volume with the source code is mounted with owner ID 65534, causing
git to throw an error and Go tools to fail. To overcome this we mark the
source code repository as safe in when the task starts.

Signed-off-by: Pavlos Tzianos <pavlos@ocurity.com>