Do not return error when remote file is fetched (Fixes #126)
This commit moves some code away from the build command, that manages
the loading of the base pipeline, task and the pipeline tasks into a
separate piece of code that lives in the same directory as the Tekton
backend. Some tests are also added to the codebase testing the correct
returning of errors. Finally, it fixes the issue where the
httpFileLoader returns an error wrapping a nil error even if the file
has been fetched from the remote source correctly.

Signed-off-by: Pavlos Tzianos <pavlos@ocurity.com>