Make HTTPFileLoader more intelligent (Fixes #128)
Splits away the Loader and its implementations into the pkg/files module
and adds more logic to the HTTPFileLoader so that it can append the
target file to the URL if it's missing. Tests are also added to make
sure that the HTTPFileLoader is 100% covered now.
A new interface and a mock implementation has been added for HTTP
requests.