fixes 154: Integrate remark linter to linter stack
This commit introduces various changes required to add remark to our
stack of linters run by reviewdog. Remark is a well-known linter for
Markdown using nodejs. It's errorformat is automatically recognised by
reviewdog making it trivial to add to the pipeline and get comments on
our PRs about potential issues.

As mentioned previously, remark is written using nodejs, hence an extra
step has been added in the lint Github action to install NodeJS on the
system and the `install-lint-tools` make target has been extended to
install dependencies listed in the package.json file. The `npm/remark`
lint runner has been added to the reviewdog stack and a .remarkignore
file is used  to configure the linter to not check the `vendor` folder
for markdown files. The linter is pluggable and we have the standard and
recommended rules installed via npm.

The format Github action has not been extended and only checks for
issues in the Go code, since the remark linter is also used as a
markdown formatter but its results will be exactly equivalent to the
ones reported (if any) by reviewdog. The user can run the `fmt-md`
target to format their Markdown files according to what remark expects.