fixes 152: Force 'migrations inspect' command to output to stdout
In the last couple of days the migrations test have failed locally for
no apparent reason, since no code related to them has changed, the
dependency versions have not changed and neither have the migrations
themselves. After a little bit of digging it looks like something in the
ubuntu container is causing the 'OutOrStderr' function call in the
migrations inspect command to return the stderr instead of stdout which
in turn causes the jq check in the migrations bash script to get an
empty input instead of the JSON doc which in turn causes it to fail
immediately. Changing the 'OutOrStderr' to 'OutOrStdout' ensures that
the code will always output to stdout as expected or whatever is the out
stream.