fixes 160: add typed k8s and Tekton V1Beta1 clientset
This new clientset is hidden behind the ClientInterface which is an
amalgamation of the core K8s clientset and the TektonV1Beta clientset.
Most importantly, it introduces the Apply method which replicates the
kubectl apply command, received a K8s runtime.Object and applying it on
the cluster using server-side apply
(https://kubernetes.io/docs/reference/using-api/server-side-apply/).

The clientset is complimented with a fake implementation that can be
used by the user to intercept the objects passed to the Apply method and
inject errors.