fixes 159: modify draconctl compiling make target to always build binary
The previous version of the draconctl compiling make target would only
compile the binary if the binary didn't exist in the expected path. This
however is a problem if the code changes, because Makefile doesn't have
the ability to track that and recompile the binary. This commit changes
the target to be a phony one so that the binary is always recompiled if
necessary. Go internally is able to track changes and produce a new
binary only when needed, so this shouldn't cause anything to slow down
but it will definitely improve the development process.