fixes 167: add draconctl command for producing Helm package of components
This command simplifies the way that we use Helm to deploy our Tasks and
pipelines by splitting away the Task management from the Pipeline
deployment. Specifically, we introduce a new draconctl command that
builds a Helm package from the components folder containing all the
Tasks discovered. The Tasks are modified to acquire anchors and whatever
else is needed to get them to work in our Pipelines. It also changes the
expected Kustomization format that can find by allowing users to
reference the Tasks deployed using a Helm package by using the release
name as the ID of a PURL.