fixes 169: pin all components Tasks to latest version
we used to bundle Tasks with Pipelines in a Helm package and we
templated the version of the component images to allow it to be set by
the user when deploying the Helm package. we no longer need that because
the version can be pinned during the creation of the one Helm package
that packages all the components in a Dracon version. we instead replace
it with 'latest'.