set component package version to original tag value
When passing the Dracon version to the `draconctl components package` in
the Github publish action, we pass the modified value of the tag that is
missing the initial 'v'. But that produces a wrong Helm package because
the Chart version is a semantic version (which is correct) but the
Dracon version and the version set in the Task images is different than
the one that the containers are pinned with, which contains the initial
'v'. This commit fixes the publish action and also does a minor fix in
the code that ensure the value given to the '--version' flag is passed
correctly all the way down to the code that constructs the Helm
package.