fix 170: add more checks for unresolved components
before deploying a pipeline we resolve all the components in two stages.
first we resolve components that are in the local filesystem or that
will be downloaded. then, in the 'Deploy' method of the orchestrator
implementations we deploy unmanaged Tasks and check that all the managed
Tasks are present in the cluster. we should have more checks that the
unmanaged Tasks have been resolved correctly before trying to do
anything with them.