180: add dracon_scan_start_time, dracon_scan_id and dracon_scan_tags to producers
when we process tasks loaded from the file system, we need to add
various results, parameters and potentially environment variables in
order for those Tasks to be integrated into our pipelines correctly.
There are two places in our codebase where this processing must happen:
in the code that creates a Helm package with all the Tasks and in the
Deploy method of the Orchestrator implementations that could deploy on
the fly Tasks. Until now the code that performed this processing was
copied in two places and wasn't explicitly tested, which is a very bad
idea. This commit fixes that by moving the code in the `pkg/components`
package and fixes an issue with the addition of the
dracon_scan_start_time, dracon_scan_id and dracon_scan tags environment
variables which are supposed to be added to the producers but where
erroneously added to the Tasks that where dicovered on the fly by the
orchestrator instead of being added during the Helm packaging process.