redirect reviewdog stderr to stdout
the lint Make target will output in stderr all the issues that
reviewdog discovered for the diff under investigation, however that
stderr is not captured by the Github action. we need to send the stderr
to stdout so that we can see it in the Github action logs when the lint
action fails.