🔥 Remove `enricher-policy` from example pipelines
Removing this because it requires 5GB of memory to run, but
Docker Desktop by default only assigns 2GB to itself. The
component can still be run, but requires users to manually
tweak their memory limit. We decided this is too complex for
an example pipeline and would cause more harm than good.