introduce multi-platform builds #334
This commit extends the base Dockerfile so that it can copy a binary
from an arbitrary path and copy it into a target one inside the
filesystem of the container image.
The scripts building the component binaries and containers have also
been extended with environment variables to help them build the binaries
for the expected architecture. The Makefile has acquired also the same
environment variable which are set by default in the host's
architecture. In the case of the `dev-deploy` target, the architecture
is set by default to `linux/amd64` since containers are usually executed
on linux/amd64 systems or on linux/amd64 VMs running the Docker daemon
(Windows, some Macs) or on VMs running linux/arm64 which supports amd64
binaries out of the box.